# shared/specs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Edge AI Platform API
  version: 1.0.0
  description: |
    REST API for the Edge AI Platform web management interface.
    Provides device management, chat, and metrics access.
  contact:
    name: VLMChat Platform
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://localhost:3000
    description: Docker development server
security:
  - bearerAuth: []
...
paths:
...
  /api/search-terms:
    get:
      tags: [search]
      summary: List all search terms
      description: Retrieve all active search terms for the current user
      operationId: listSearchTerms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of search terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchTerm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [search]
      summary: Create search term
      description: |
        Create new search term from natural language query.
        Server parses via Groq API to generate filter vectors,
        saves to database, and broadcasts MQTT filter update to all devices.
      operationId: createSearchTerm
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [searchString]
              properties:
                searchString:
                  type: string
                  description: Natural language search query
                  minLength: 1
                  maxLength: 500
                  example: "person wearing red shirt"
      responses:
        '201':
          description: Search term created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTerm'
        '400':
          description: Invalid search string or parsing failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/search-terms/{id}:
    get:
      tags: [search]
      summary: Get search term details
      description: Retrieve specific search term with full filter configuration
      operationId: getSearchTerm
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Search term ID
      responses:
        '200':
          description: Search term details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTerm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [search]
      summary: Delete search term
      description: |
        Remove search term from database and broadcast updated MQTT filter
        to all devices (without this term).
      operationId: deleteSearchTerm
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Search term ID
      responses:
        '200':
          description: Search term deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    SearchTerm:
      description: User-defined search term with parsed filter configuration
      allOf:
        - $ref: '../schemas/search-term-v1.0.0.json'
...
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from cookie-based authentication